Backfill, disable ehcache to reduce OldGen use.
